feat: refactor in roles permissions tabs
